requirements.txt
certifi==2019.6.16
chardet==3.0.4
coverage==4.5.4
Django==2.2.4
django-filter==2.2.0
djangorestframework==3.10.1
idna==2.8
Markdown==3.1.1
psycopg2==2.8.3
psycopg2-binary==2.8.3
python-memcached==1.59
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
